requirements.txt
Django==1.11.5
pytz==2017.2
Django==1.11.5
djangorestframework==3.6.4
pytz==2017.2
